Extract schema compiler into lightweight ihp-schema-compiler package
Production builds require ihp-ide just to run build-generated-code, which
compiles Schema.sql into Types.hs. But ihp-ide has 91+ dependencies making
it slow to build.

This creates a new ihp-schema-compiler package containing only what's needed
for schema compilation (~15 dependencies), significantly reducing build times
for production deployments.

Changes:
- Create ihp-schema-compiler package with IHP.SchemaCompiler and
  IHP.SchemaCompiler.Parser modules
- Move build-generated-code executable to ihp-schema-compiler
- Update ihp-ide to depend on ihp-schema-compiler
- Update imports in ihp-ide to use new module paths
- Update Nix configuration to use ihp-schema-compiler for builds
