+++ ihp-ide/IHP/IDE/CodeGen/MailGenerator.hs
@@ -3,7 +3,7 @@ module IHP.IDE.CodeGen.MailGenerator (buildPlan, buildPlan', MailConfig (..)) wh
 import IHP.Prelude
 import IHP.IDE.CodeGen.Types
 import qualified Data.Text as Text
-import qualified IHP.IDE.SchemaDesigner.Parser as SchemaDesigner
+import qualified IHP.SchemaCompiler.Parser as SchemaDesigner
 import IHP.Postgres.Types
 import Text.Countable (singularize, pluralize)
