+++ ihp-ide/IHP/IDE/SchemaDesigner/Controller/Schema.hs
@@ -7,7 +7,7 @@ import IHP.IDE.SchemaDesigner.View.Schema.Code
 import IHP.IDE.SchemaDesigner.View.Schema.GeneratedCode
 import IHP.IDE.SchemaDesigner.View.Schema.SchemaUpdateFailed
 
-import IHP.IDE.SchemaDesigner.Parser
+import IHP.SchemaCompiler.Parser
 import qualified IHP.SchemaCompiler as SchemaCompiler
 import qualified System.Process as Process
 import System.Exit
