+++ ihp-schema-compiler/ihp-schema-compiler.cabal
@@ -1,0 +1,89 @@
+cabal-version:       2.2
+name:                ihp-schema-compiler
+version:             1.4.0
+synopsis:            Lightweight schema compiler for IHP
+description:         Compiles Schema.sql to Generated Types.hs for IHP applications. This is a minimal package with few dependencies, suitable for production builds that don't need the full ihp-ide.
+license:             MIT
+author:              digitally induced GmbH
+maintainer:          hello@digitallyinduced.com
+homepage:            https://ihp.digitallyinduced.com/
+bug-reports:         https://github.com/digitallyinduced/ihp/issues
+copyright:           (c) digitally induced GmbH
+category:            Web, IHP
+stability:           Stable
+tested-with:         GHC == 9.8.4
+build-type:          Simple
+data-dir: data
+data-files: IHPSchema.sql
+
+Flag FastBuild
+    Default: False
+    Description: Disables all optimisations, leads to faster build time
+
+common shared-properties
+    default-language: GHC2021
+    build-depends:
+            base >= 4.17.0 && < 4.22
+          , classy-prelude
+          , text
+          , bytestring
+          , directory
+          , string-conversions
+          , interpolate
+          , split
+          , neat-interpolation
+          , countable-inflections
+          , with-utf8
+          , ihp
+          , ihp-postgres-parser
+    default-extensions:
+        OverloadedStrings
+        , NoImplicitPrelude
+        , ImplicitParams
+        , DisambiguateRecordFields
+        , DuplicateRecordFields
+        , OverloadedLabels
+        , DataKinds
+        , QuasiQuotes
+        , TypeFamilies
+        , PackageImports
+        , RecordWildCards
+        , BlockArguments
+        , LambdaCase
+        , OverloadedRecordDot
+    if flag(FastBuild)
+        ghc-options:
+            -Wunused-imports
+            -Wunused-foralls
+            -Wmissing-fields
+            -Winaccessible-code
+            -Wno-ambiguous-fields
+    else
+        ghc-options:
+            -fstatic-argument-transformation
+            -funbox-strict-fields
+            -haddock
+
+            -Wunused-imports
+            -Wunused-foralls
+            -Wmissing-fields
+            -Winaccessible-code
+            -fspecialise-aggressively
+            -Wno-ambiguous-fields
+
+library
+    import: shared-properties
+    hs-source-dirs: .
+    exposed-modules:
+          IHP.SchemaCompiler
+        , IHP.SchemaCompiler.Parser
+
+executable build-generated-code
+    import: shared-properties
+    build-depends: ihp-schema-compiler
+    hs-source-dirs: exe
+    main-is: IHP/CLI/BuildGeneratedCode.hs
+
+source-repository head
+    type:     git
+    location: https://github.com/digitallyinduced/ihp
